refactor: Break down DetailedMetricsPanel into smaller components
- Create reusable CollapsibleCard component for all expandable sections
- Extract specialized card components for different metric types
- Implement CategoryScoresPanel and InsightsPanel as separate components
- Add conditional debug logging with DEBUG_MODE flag (to be generalized in future updates)
- Improve handling of placeholder data with clear labeling
- Update README.md to reflect new component structure
- Fix TypeScript type issues with icon names
